category and sub category model fixing

+++ src/core/database/migrations/20220115085123-create_category.js
@@ -26,22 +26,22 @@ module.exports = {
         },
 
         image_id: {
-          type: Sequelize.BIGINT,
+          type: Sequelize.INTEGER,
           allowNull: false,
         },
 
         is_sub_category: {
-          type: Sequelize.BIGINT,
+          type: Sequelize.INTEGER,
           allowNull: false,
         },
 
         is_active: {
-          type: Sequelize.BIGINT,
+          type: Sequelize.INTEGER,
           allowNull: false,
         },
 
         is_deleted: {
-          type: Sequelize.BIGINT,
+          type: Sequelize.INTEGER,
           allowNull: false,
           defaultValue: 0,
         },
